microarray: refactor 'xrorthologs.py' to move into bcftbx.cli.
