+++ bin/xrorthologs.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     xrorthologs.py: cross-reference data for two species given probeset lookup
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.xrorthologs import main
+if __name__ == "__main__":
+     main()
